refactor(github): split monolithic client.go into domain-specific files
